push rope back to encoder/decoder
