+++ eole/predict/inference.py
@@ -630,7 +630,6 @@ class Inference(object):
             emb = self.model.tgt_emb(decoder_in, step=step)
 
         tgt_pad_mask = decoder_in.eq(self._tgt_pad_idx).unsqueeze(1)  # [B, 1, T_tgt]
-        position_embeddings = self.model.rope.update(decoder_in.size(1), step=step)
         dec_out, dec_attn = self.model.decoder(
             emb,
             enc_out=enc_out,
@@ -639,7 +638,6 @@ class Inference(object):
             return_attn=self.global_scorer.has_cov_pen or return_attn,
             src_pad_mask=src_pad_mask,
             tgt_pad_mask=tgt_pad_mask,
-            position_embeddings=position_embeddings,
             left_pad=left_pad,
         )
         # Generator forward.
