+++ eole/tests/data/data_lm/gen-nucleus-sampling-sol1.txt
@@ -1,7 +1,7 @@
-, ?
-in German Presidency in German Presidency programmes first cosy apartment in Slovene style .
-higher in New Christ .
-"
-fine figure on serving angles and the fine words 20ies last century Without density properly prepared to strive as soon as easy as a enterprise events to fallen into procurement procedures corresponding increasingly difficult it .
-have any chance , staggered according to the requirement or reproductive schools .
-serious , the giant .
+you ! It is !
+in the German German Presidency in southern district .
+the top of the top of the future .
+Halle lent money he caused by debt-laden countries where they went at LKW Walter International AG and Bond Street London Underground Station .
+fine words have underlined the formation ...
+happens so . In 1971 , Warren de Warren Rudman , then the Attorney General of New Hampshire , selected him to be recognized .
+80 km from the " s of Munich airport .
